+++ tasks/gulp/compile-assets.mjs
@@ -1,23 +1,17 @@
 
-import { join, parse } from 'path'
+import { join } from 'path'
 
 import gulp from 'gulp'
-import rollup from 'gulp-better-rollup'
-import gulpif from 'gulp-if'
 import plumber from 'gulp-plumber'
 import postcss from 'gulp-postcss'
 import rename from 'gulp-rename'
 import gulpSass from 'gulp-sass'
-import uglify from 'gulp-uglify'
 import merge from 'merge-stream'
-import minimatch from 'minimatch'
 import nodeSass from 'node-sass'
 import PluginError from 'plugin-error'
 import slash from 'slash'
 
 import { paths, pkg } from '../../config/index.js'
-import { getListing } from '../../lib/file-helper.js'
-import { componentNameToJavaScriptModuleName } from '../../lib/helper-functions.js'
 import { destination, isDev, isDist, isPackage } from '../task-arguments.mjs'
 
 const sass = gulpSass(nodeSass)
@@ -109,86 +103,6 @@ function compileStylesheet (stream, options = {}) {
 }
 
 /**
- * Compile JavaScript ESM to CommonJS task
- *
- * @returns {Promise<import('stream').Stream>} Output file stream
- */
-export async function compileJavaScripts () {
-  const destPath = isPackage
-    ? join(destination, 'govuk')
-    : destination
-
-  // For dist/ folder we only want compiled 'all.js'
-  const fileLookup = isDist ? 'govuk/all.mjs' : 'govuk/**/!(*.test).mjs'
-
-  // Perform a search and return an array of matching file names
-  const srcFiles = await getListing(paths.src, fileLookup)
-
-  return merge(srcFiles.map((file) => {
-    const { dir: srcPath, name } = parse(file)
-
-    // This is combined with destPath in gulp.dest()
-    // so the files are output to the correct folders
-    const modulePath = slash(srcPath).replace(/^govuk\/?/, '')
-
-    // We only want to give component JavaScript a unique module name
-    const moduleName = minimatch(srcPath, '**/components/**')
-      ? componentNameToJavaScriptModuleName(name)
-      : 'GOVUKFrontend'
-
-    return compileJavaScript(gulp.src(slash(join(paths.src, file)), {
-      sourcemaps: true
-    }), moduleName)
-      .pipe(gulp.dest(slash(join(destPath, modulePath)), {
-        sourcemaps: '.'
-      }))
-  }))
-}
-
-compileJavaScripts.displayName = 'compile:js'
-
-/**
- * Compile JavaScript ESM to CommonJS helper
- *
- * @param {import('stream').Stream} stream - Input file stream
- * @param {string} moduleName - Rollup module name
- * @returns {import('stream').Stream} Output file stream
- */
-function compileJavaScript (stream, moduleName) {
-  return stream
-    .pipe(plumber(errorHandler(stream, 'compile:js')))
-
-    // Compile JavaScript ESM to CommonJS
-    .pipe(rollup({
-      // Used to set the `window` global and UMD/AMD export name
-      // Component JavaScript is given a unique name to aid individual imports, e.g GOVUKFrontend.Accordion
-      name: moduleName,
-      // Legacy mode is required for IE8 support
-      legacy: true,
-      // UMD allows the published bundle to work in CommonJS and in the browser.
-      format: 'umd'
-    }))
-
-    // Minify
-    .pipe(gulpif(isDist, uglify({
-      ie8: true
-    })))
-
-    .pipe(plumber.stop())
-
-    // Rename
-    .pipe(gulpif(isDist,
-      rename({
-        basename: 'govuk-frontend',
-        suffix: `-${pkg.version}.min`
-      })
-    ))
-    .pipe(rename({
-      extname: '.js'
-    }))
-}
-
-/**
  * Compiler error handler
  *
  * Sets up Gulp plumber to suppress errors during development
@@ -198,7 +112,7 @@ function compileJavaScript (stream, moduleName) {
  * @param {string} name - Task name for non-plugin error logging
  * @returns {import('stream').Stream} Output file stream
  */
-function errorHandler (stream, name) {
+export function errorHandler (stream, name) {
   return function (cause) {
     const error = new PluginError(cause.plugin || name, cause.messageFormatted || cause)
